requirements.txt
streamlit
streamlit-drawable-canvas
scikit-learn
numpy
Pillow
joblib
